tradingagents-cc/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
yfinance>=0.2.36
pandas-ta>=0.3.14b
textblob>=0.17.1
vaderSentiment>=3.3.2
newsapi-python>=0.2.7
duckduckgo-search>=6.0.0
mcp>=1.0.0
PyYAML>=6.0
pytest>=7.4.0
pyarrow>=10.0.0
alpaca-trade-api>=3.0.0
ib_insync>=0.9.86
praw>=7.7.0
requests>=2.31.0
python-dotenv>=1.0.0
